refactor: 기본 엔티티 클래스 패키지 구조 변경 (#45)
